requirements.txt
certifi>=2024.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0
slack-sdk>=3.30.0
slack-bolt>=1.18.0
requests>=2.31.0
substack-api>=1.0.0
python-substack>=0.1.18
schedule>=1.2.0
